pyproject.toml
[tool.poetry]
name = "nonebot_plugin_auto_teenstudy"
version = "0.1.0"
description = "基于nonebot异步框架的青年大学自动提交插件"
authors = ["ZMXC01 <2393899036@qq.com>"]
license="MIT"
readme="README.md"
homepage="https://github.com/ZMXC01/nonebot_plugin_auto_teenstudy"
[tool.poetry.dependencies]
python = "^3.6"
anti_useragent="~=1.0.7"
nonebot2 = ">=2.0.0b2"
nonebot-adapter-onebot = "^2.0.0b1"
Pillow="^~=9.1.0"
beautifulsoup4 = "~=4.10.0"
httpx="~=0.20.0"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
